Remove external uses of `Vfs`
