+++ compiler/toc_analysis/src/test_db.rs
@@ -8,7 +8,7 @@ use toc_ast_db::{
 };
 use toc_hir::library::LibraryId;
 use toc_salsa::salsa;
-use toc_vfs_db::db::VfsDatabaseExt;
+use toc_vfs_db::db::{PathIntern, VfsDatabaseExt};
 
 #[salsa::database(
     toc_vfs_db::db::FileSystemStorage,
@@ -23,7 +23,6 @@ use toc_vfs_db::db::VfsDatabaseExt;
 #[derive(Default)]
 pub(crate) struct TestDb {
     storage: salsa::Storage<Self>,
-    vfs: toc_vfs::Vfs,
 }
 
 impl salsa::Database for TestDb {}
@@ -34,7 +33,7 @@ impl TestDb {
         let fixture = toc_vfs::generate_vfs(source).unwrap();
         db.insert_fixture(fixture);
 
-        let root_file = db.vfs.intern_path("src/main.t".into());
+        let root_file = db.intern_path("src/main.t".into());
         let mut source_graph = SourceGraph::default();
         let library_id = source_graph.add_library(toc_hir::library_graph::SourceLibrary {
             name: "main".into(),
