+++ compiler/toc_driver/src/main.rs
@@ -12,7 +12,7 @@ use toc_hir_db::db::HirDatabase;
 use toc_salsa::salsa;
 use toc_span::{FileId, Span};
 use toc_vfs::FileLoader;
-use toc_vfs_db::db::FileSystem;
+use toc_vfs_db::db::{FileSystem, PathIntern};
 
 mod config;
 
@@ -41,7 +41,7 @@ fn main() {
     let mut db = MainDatabase::default();
 
     // Add the root path to the db
-    let root_file = db.vfs.intern_path(path);
+    let root_file = db.intern_path(path.try_into().unwrap());
 
     // Set the source root
     let mut source_graph = SourceGraph::default();
@@ -230,7 +230,6 @@ impl ariadne::Cache<FileId> for VfsCache<'_> {
 #[derive(Default)]
 struct MainDatabase {
     storage: salsa::Storage<Self>,
-    vfs: toc_vfs::Vfs,
 }
 
 impl salsa::Database for MainDatabase {}
