+++ lsp-server/toc-ide-db/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 lsp-types = "0.93.2"
 serde_json = "1.0.64"
 tracing = { workspace = true }
+camino = { workspace = true }
 ropey = "1.5.0"
 
 # Local compiler dependencies
